// src/app/models/Agricultor.ts
import { Users } from "./Users";

export class Agricultor{
    idAgricultor: number = 0;
    nombreAgricultor: string = "";
    correoAgricultor: string = "";
    telefonoAgricultor: string = "";
    ubicacionAgricultor: string = "";
    tipoProductoAgricultor: string = "";
    usuario: Users = new Users();
}